map_type fixes. mapnik->standard, hide osmarender
The map_type 'mapnik' value is badly named. Create a new alias 'standard'
which is now the new default (which the old 'mapnik' value continues to work as
a deprecated value)

'osmarender' has been (badly) mapped to MapQuest open tiles for a while now, so
broken because these were discontinued (as was the original 'osmarender' layer)
Fix maptype=osmarender so it works, by mapping to osm standard, but obviously
this value is also deprecated.

Changes to the intro page to reflect this.

Fixes https://github.com/dfacts/staticmaplite/issues/3